Connect Explore child components to store
